pyproject.toml
[tool.poetry]
name = "dokku-updater"
version = "0.1.0"
description = "Automatically deploy dokku from docker registry"
authors = ["Jeong Arm"]

[tool.poetry.dependencies]
python = "^3.8"
docker = "^5.0.0"
PyYAML = "^5.4.1"
APScheduler = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
